ref(escalating-issues): Move functions to centralize helper functions
